% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIERCE.R
\name{build_adata_object}
\alias{build_adata_object}
\title{build_adata_object}
\usage{
build_adata_object(
  loom_file = NULL,
  adata_object = NULL,
  Seurat_object = NULL,
  replace_matrix = FALSE,
  matrix_slot = "counts",
  subset_key_Seurat = NULL,
  subset_ident_Seurat = NULL,
  active_assay = NULL,
  Seurat_cellnames_prefix = NULL,
  Seurat_cellnames_suffix = NULL,
  sample_column = NULL,
  metadata_keys = NULL,
  reduction_keys = NULL,
  subset_key_anndata = NULL,
  subset_ident_anndata = NULL,
  cells_to_keep = NULL,
  genes_to_keep = NULL,
  user_metadata = NULL,
  user_embedding = NULL,
  user_embedding_name = NULL
)
}
\arguments{
\item{loom_file}{name of the loom file generated by velocyto (including its path). More than one file can be specified with a character vector; in this case, the new anndata object will include all the samples merged together}

\item{adata_object}{alternatively to creating a brand new anndata object, a pre-existing anndata object can be provided and modified through the options below. Useful for modifying any anndata object on the fly}

\item{Seurat_object}{optional Seurat object to be integrated with the new anndata object. WARNING: if a Seurat object is provided, only the cells and the genes stored in its active assay will be kept in the new anndata object, the remaining will be discarded}

\item{replace_matrix}{boolean; if TRUE, the expression matrix stored in the active assay of the provided Seurat object will be stored in the "X" slot of the anndata object; if FALSE (default), the spliced counts from the loom file will be used instead}

\item{matrix_slot}{only used if replace_matrix=TRUE; slot of the active assay of the Seurat object from which the expression matrix has to be retrieved ("counts" by default, set to "data" if interested in the normalized counts)}

\item{subset_key_Seurat}{optional; name of the column of the "meta.data" slot to consider if the user is exclusively interested in a particular subset of the Seurat object (see subset_ident_Seurat)}

\item{subset_ident_Seurat}{character vector containing the cell identities to keep in the subset of the Seurat object, amongst those included in the column specified in subset_key_Seurat}

\item{active_assay}{name of the active assay of the Seurat object; by default, the current active assay will be considered}

\item{Seurat_cellnames_prefix}{character vector containing the complete list of all the unique strings preceding the barcodes in the cell names of the Seurat object; necessary to adapt the nomenclature of the Seurat object to the nomenclature of the anndata object. WARNING: each unique string must be specified only once, but all strings need to be exact, an error will be thrown otherwise. This parameter does not need to be specified if the barcodes are not preceded by any string}

\item{Seurat_cellnames_suffix}{character vector containing the complete list of all the unique strings following the barcodes in the cell names of the Seurat object; necessary to adapt the nomenclature of the Seurat object to the nomenclature of the anndata object. WARNING: each unique string must be specified only once, but all strings need to be exact, an error will be thrown otherwise. This parameter does not need to be specified if the barcodes are not followed by any string}

\item{sample_column}{name of the column of the "meta.data" slot of the Seurat object containing the sample identity of each cell; necessary to adapt the nomenclature of the Seurat object to the nomenclature of the anndata object. WARNING: for each single cell, the sample identity MUST correspond to the name of the respective loom file (without the ".loom" extension), an error will be thrown otherwise}

\item{metadata_keys}{character vector containing the list of the columns of the "meta.data" slot of the Seurat object (cell annotations) that need to be transferred to the "obs" slot of the new anndata object}

\item{reduction_keys}{character vector containing the list of the dimension reductions of the Seurat object that need to be transferred to the "obsm" slot of the new anndata object}

\item{subset_key_anndata}{optional; name of the cell annotation in the "obs" slot to consider if the user is exclusively interested in a particular subset of the anndata object (see subset_ident_anndata)}

\item{subset_ident_anndata}{character vector containing the cell identities to keep in the subset of the anndata object, amongst those that are included in the annotation specified in subset_key_anndata}

\item{cells_to_keep}{optional character vector containing the names of the cells to exclusively include in the new anndata object. Cell names need to be in loom format: <sample_name>:<cell_barcode>x}

\item{genes_to_keep}{optional character vector containing the names of the genes to exclusively include in the new anndata object}

\item{user_metadata}{optional user-provided data frame containing additional cell annotations to include in the "obs" slot of the new anndata object. The row names of the data frame must correspond exactly to the cell names ("obs_names" slot) of the new anndata object (as well as to the cell names of the active assay of the Seurat object if provided). Also, they need to be in loom format: <sample_name>:<cell_barcode>x}

\item{user_embedding}{optional user-provided data frame containing an additional dimension reduction to include in the "obsm" slot of new anndata object. The data frame must include a separate column for each dimension, and the row names must correspond exactly to the cell names ("obs_names" slot) of the new anndata object (as well as to the cell names of the active assay of the Seurat object if provided)}

\item{user_embedding_name}{name of the additional dimension reduction provided in user_embedding}
}
\value{
The new anndata object, optionally integrated with the user-provided Seurat object, annotations and dimension reductions
}
\description{
Creates a new anndata object from a loom file generated by velocyto. Spliced and unspliced read counts will be stored into separate layers. If not provided by the user, the gene expression matrix (slot "X") of the anndata object will correspond to spliced counts. Optionally, the anndata object can be integrated with a Seurat object, to retrieve useful information such as cell annotations and dimension reductions; these can also be directly provided by the user as separate data frames
}
\examples{
#Simply generate an anndata object from a single loom file with no additional operations

adata <- build_adata_object("./sample_name/sample_name.loom")

#Generate a merged anndata object from two separate loom files and integrate it with a Seurat object. The latter includes multiple samples merged together, but ONLY two of them (Sample1 and Sample2) will be included in the new anndata object. Note that the column of the "meta.data" slot specified in sample_column and the names of the loom files MUST follow the same nomenclature. Several cell annotations and dimension reductions are retrieved from the Seurat object. The unique strings preceding and following the barcodes in the cell names of the Seurat object are accurately specified to allow the integration with the anndata object

adata <- build_adata_object(loom_file=c("./Sample1/Sample1.loom", "./Sample2/Sample2.loom"), Seurat_object=name_of_Seurat_object, replace_matrix=TRUE, matrix_slot="counts", subset_key_Seurat="Sample", subset_ident_Seurat=c("Sample1", "Sample2"), Seurat_cellnames_prefix=c("Sample1_", "Sample2_"), Seurat_cellnames_suffix=c("-1"), sample_column="Sample", metadata_keys=c("Sample", "Clusters", "Cell_cycle_phase"), reduction_keys=c("pca", "umap"))

#Generate an anndata object from a single loom file and include additional user-provided cell annotations and dimension reductions. Only the cells and the genes specified in the respective character vectors are included. Note that the row names of the additional data frames provided by the user MUST correspond exactly to the cell names specified in cells_to_keep

adata <- build_adata_object(loom_file="./sample_name/sample_name.loom", cells_to_keep=cell_names_vector, genes_to_keep=gene_names_vector, user_metadata=annotation_data_frame, user_embedding=dimension_reduction_data_frame, user_embedding_name="new_dimension_reduction")

}
