% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIERCE.R
\name{compute_future_states}
\alias{compute_future_states}
\title{compute_future_states}
\usage{
compute_future_states(
  adata,
  min_counts = NULL,
  min_cells = NULL,
  max_counts = NULL,
  max_cells = NULL,
  min_counts_u = NULL,
  min_cells_u = NULL,
  max_counts_u = NULL,
  max_cells_u = NULL,
  min_shared_counts = NULL,
  min_shared_cells = NULL,
  retain_genes = NULL,
  n_pcs = NULL,
  n_neighbors = 30,
  mode_moments = "connectivities",
  use_raw = FALSE,
  n_cores = NULL,
  adata_copy = FALSE
)
}
\arguments{
\item{adata}{anndata object with spliced and unspliced gene counts stored into the "spliced" and "unspliced" layers, respectively}

\item{min_counts}{minimum number of spliced counts required for a gene to be retained for velocity computation}

\item{min_cells}{minimum number of expressing cells (spliced counts only) required for a gene to be retained for velocity computation}

\item{max_counts}{maximum number of spliced counts allowed for a gene to be retained for velocity computation}

\item{max_cells}{maximum number of expressing cells (spliced counts only) allowed for a gene to be retained for velocity computation}

\item{min_counts_u}{minimum number of unspliced counts required for a gene to be retained for velocity computation}

\item{min_cells_u}{maximum number of unspliced counts allowed for a gene to be retained for velocity computation}

\item{max_counts_u}{minimum number of expressing cells (unspliced counts only) required for a gene to be retained for velocity computation}

\item{max_cells_u}{maximum number of expressing cells (unspliced counts only) allowed for a gene to be retained for velocity computation}

\item{min_shared_counts}{minimum number of spliced OR unspliced counts required for a gene to be retained for velocity computation}

\item{min_shared_cells}{minimum number of expressing cells (spliced OR unspliced counts) required for a gene to be retained for velocity computation}

\item{retain_genes}{character vector containing a list of genes to be retained for velocity computation regardless of the filters}

\item{n_pcs}{number of PCs to consider for the nearest neighbor graph computation (necessary for first order moments computation, skipped if use_raw=TRUE, or if the graph is already present)}

\item{n_neighbors}{number of nearest neighbors to compute for each cell in the nearest neighbor graph (necessary for first order moments computation, skipped if use_raw=TRUE, or if the graph is already present). Default is 30 neighbors}

\item{mode_moments}{method for first order moments computation. Choose between "distances" or "connectivities". Default is "connectivities"}

\item{use_raw}{boolean; whether to use normalized counts for velocity computation instead of first order moments. Default is FALSE}

\item{n_cores}{number of cores to use for dynamical model fitting. If NULL (default), only 1 will be used}

\item{adata_copy}{boolean; if TRUE, create a new anndata object with the results of the analysis; if FALSE, update the provided anndata object. Default is FALSE}
}
\value{
If adata_copy=TRUE, a new anndata object with the results of the analysis will be returned. Otherwise (default), the provided anndata object will be updated with the results of the analysis. In both cases, velocities, first order moments (both spliced and unspliced) and future spliced transcriptional states will be saved as additional layers
}
\description{
Computes gene velocities with scVelo and infers future spliced transcriptional states by summing velocities to observed states. Velocities can be computed either from the spliced and unspliced normalized counts, or from the spliced and unspliced first order moments (default procedure of scVelo). The spliced transcriptional states will be consequently defined either as spliced normalized counts, or spliced first order moments
}
\examples{
#Compute velocities and future spliced transcriptional states for all genes

compute_future_states(adata)

#Compute velocities and future spliced transcriptional states only for the genes that are expressed with at least 20 normalized spliced or unspliced counts across all cells

compute_future_states(adata, min_shared_counts=20)

#Compute velocities and future spliced transcriptional states for all genes using normalized counts instead of first order moments. Save results in a separate anndata object

adata_future_states <- compute_future_states(adata, use_raw=TRUE, adata_copy=TRUE)

}
