% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIERCE.R
\name{compute_graph_and_stream}
\alias{compute_graph_and_stream}
\title{compute_graph_and_stream}
\usage{
compute_graph_and_stream(
  adata,
  project_dir = "./Velocity_of_the_entropy_pipeline",
  only_velocity_genes = FALSE,
  n_neighbors_graph = NULL,
  n_pcs_graph = NULL,
  n_neighbors_emb = NULL,
  n_pcs_emb = NULL,
  sqrt_transform = TRUE,
  embedding_basis = "umap_entropy",
  force_graph_recalc = TRUE,
  color_as = "total_entropies_observed",
  lab_order = NULL,
  palette = NULL,
  legend_loc = "right margin",
  alpha = 0.3,
  add_outline = TRUE,
  min_mass = 4,
  n_cores = NULL,
  adata_copy = TRUE
)
}
\arguments{
\item{adata}{anndata object with "partial_entropies_observed" and "velocity_of_the_entropy" layers (produced by the "compute_signaling_entropy" function)}

\item{project_dir}{name of the directory containing the results of the main FIERCE analysis (including the path). If it does not exist, it will be created. The default name is "./Velocity_of_the_entropy_pipeline". The streamplot will be saved in the "velocity_field_streamplots" sub-directory}

\item{only_velocity_genes}{boolean; whether only the genes that successfully fitted the velocity dynamical model ("compute_velocity" function) should be used for transition probabilities computation. Default is FALSE}

\item{n_neighbors_graph}{this parameter, in combination with "n_pcs_graph", specifies which nearest neighbors graph will be used to compute the cell-cell transition probabilities on the entropy space. By default, the last graph computed by the "compute_entropy_UMAP" function will be used}

\item{n_pcs_graph}{this parameter, in combination with "n_neighbors_graph", specifies which nearest neighbors graph will be used to compute the cell-cell transition probabilities on the entropy space. By default, the last graph computed by the "compute_entropy_UMAP" function will be used}

\item{n_neighbors_emb}{if "umap_entropy" is chosen as cell embedding for the streamplot, this parameter, in combination with "n_pcs_emb", specifies which embedding will be used among those that have been produced by the "compute_entropy_UMAP" function (stored in the "obsm" slot). If NULL (default), the last computed embedding will be used}

\item{n_pcs_emb}{if "umap_entropy" is chosen as cell embedding for the streamplot, this parameter, in combination with "n_neighbors_emb", specifies which embedding will be used among those that have been produced by the "compute_entropy_UMAP" function (stored in the "obsm" slot). If NULL (default), the last computed embedding will be used}

\item{sqrt_transform}{boolean; whether to apply the variance-stabilizing transformation during transition probabilities computation. It helps to obtain a smoother streamplot. Default is TRUE}

\item{embedding_basis}{name of the embedding to use for the streamplot. Default is "umap_entropy"}

\item{force_graph_recalc}{whether to force the recalculation of the cell-cell transition probabilities matrix, if already present. Default is TRUE}

\item{color_as}{character vector containing the names of the cell annotations in the "obs" slot to visualize on the streamplot}

\item{lab_order}{optional; list containing, for each cell annotation specified in "color_as", a character vector specifying the order by which the respective labels should be printed in the legends of the streamplots. If it is not necessary to specify a particular order for a specific annotation, just write NULL. If it is not necessary to specify any order for any annotation, do not change the default value of this parameter}

\item{palette}{optional; list containing, for each cell annotation specified in "color_as", a character vector specifying the colors of the respective labels. If it is not necessary to specify any color for a specific annotation, just write NULL; in this case, a default color palette will be used. If it is not necessary to specify any color for any annotation, do not change the default value of this parameter; in this case, annotations will be colored according either to a default palette, or to the colors that are already stored in the anndata object ("uns" slot)}

\item{legend_loc}{localization of the legend on the streamplot. Default is "right margin"}

\item{alpha}{transparency of the dots on the streamplots, ranging from 0 (fully transparent) to 1 (fully opaque). Default is 0.3}

\item{add_outline}{whether to add an outline around groups of dots on the streamplots. Default is TRUE}

\item{min_mass}{minimum magnitude required for velocity vectors to be drawn on the streamplot, ranging from 0 (all vectors) to 5 (most intense vectors only). Default is 4}

\item{n_cores}{number of cores to use to compute the cell-cell transition probability matrix (most computationally expensive step). If NULL (default), only 1 will be used}

\item{adata_copy}{boolean; if TRUE, create a new anndata object with the results of the analysis; if FALSE, update the provided anndata object. Default is TRUE}
}
\value{
If adata_copy=TRUE (default), a new anndata object with the results of the analysis will be returned. Otherwise, the provided anndata object will be updated with the results of the analysis. In both cases, the velocity graph (correlations used for transition probabilities computation) will be saved in the "uns" slot, and the predicted displacements of cells (used for streamplot drawing) will be saved in the "obsm" slot
}
\description{
Computes the cell-cell transition probabilites based on the correlation of the velocities of the entropy of each cell with the entropies of neighboring cells. Then, based on such probabilites, draws the streamplot on an embedding of choice (the default is UMAP). scVelo has been adapted to perform all the computation
}
\examples{
#Draw the streamplot on the UMAP embedding by using all genes for transition probabilities computation. The last entropy-based embedding computed by the "compute_entropy_UMAP" function is used. In this example, the clusters, the cell types and the cell cycle phase are visualized on the streamplot. A specific label order and a specific color for each label is specified for the cell cycle phase only. The results are saved in a new anndata object

adata_velocity_entropy_stream <- compute_graph_and_stream(adata_velocity_entropy, color_as=c("clusters", "cell_types", "phase"), lab_order=list(NULL, NULL, c("G1","S","G2M")), palette=list(NULL, NULL, c("green", "red", "blue")))

#Draw the streamplot on the UMAP embedding by using only the best fit velocity genes for transition probabilities computation, and without the variance stabilizing transformation. The last entropy-based embedding computed by the "compute_entropy_UMAP" function is used. In this example, the clusters, the cell types and the cell cycle phase are visualized on the streamplot. A specific label order and a specific color for each label is specified for the cell cycle phase only. The results are saved in a new anndata object

adata_velocity_entropy_stream <- compute_graph_and_stream(adata_velocity_entropy, only_velocity_genes=TRUE, sqrt_transform=FALSE, color_as=c("clusters", "cell_types", "phase"), lab_order=list(NULL, NULL, c("G1","S","G2M")), palette=list(NULL, NULL, c("green", "red", "blue")))

#Draw the streamplot on the UMAP embedding by using all genes for transition probabilities computation. A specific entropy-derived UMAP embedding (30 neighbors and 30 PCs) is used. In this example, the clusters, the cell types and the cell cycle phase are visualized on the streamplot. A specific label order and a specific color for each label is specified for the cell cycle phase only. The provided anndata object is updated

compute_graph_and_stream(adata_velocity_entropy, n_neighbors_emb=30, n_pcs_emb=30, color_as=c("clusters", "cell_types", "phase"), lab_order=list(NULL, NULL, c("G1","S","G2M")), palette=list(NULL, NULL, c("green", "red", "blue")), adata_copy=FALSE)

}
