\name{FIERCE}
\alias{FIERCE}
\alias{FIERCE}
\docType{package}
\title{
  \packageTitle{FIERCE}
}
\description{

  \packageDescription{FIERCE}


}
\details{

  \bold{Functions Index:}
  \itemize{
    \item \code{\link{build_adata_object}}
    \item \code{\link{compute_entropy_UMAP}}
    \item \code{\link{compute_graph_and_stream}}
    \item \code{\link{compute_SCENT_entropy}}
    \item \code{\link{compute_velocity}}
    \item \code{\link{load_h5ad}}
    \item \code{\link{load_test_dataset}}
    \item \code{\link{save_h5ad}}
    \item \code{\link{perform_preprocessing}}
    \item \code{\link{plot_SCENT_results}}
  }

}
\author{
  \packageAuthor{FIERCE}

  \bold{Maintainers}:
    \itemize{
      \item Luca Calderoni \email{luca.calderoni@unimore.it}
      \item Francesco Grandi \email{francesco.grandi1993@unimore.it}
    }
}

\references{

}
\keyword{ FIERCE, scRNA-seq, Seurat, scanpy, scvelo, SCENT, entropy }
\seealso{
  Useful links:
    \itemize{
      \item \url{https://github.com/bicciatolab/FIERCE}

      \item Report bugs at \url{https://github.com/bicciatolab/FIERCE/issues}
    }
}
\examples{
}
