% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIERCE.R
\name{load_h5ad}
\alias{load_h5ad}
\title{load_h5ad}
\usage{
load_h5ad(file_name)
}
\arguments{
\item{file_name}{name of h5ad file (including its path)}
}
\value{
anndata object contained in the h5ad file
}
\description{
Loads anndata object from h5ad file
}
\examples{
adata <- load_h5ad("./anndata_object.h5ad")

}
