% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIERCE.R
\name{load_test_dataset}
\alias{load_test_dataset}
\title{load_test_dataset}
\usage{
load_test_dataset(dataset)
}
\arguments{
\item{dataset}{name of the test dataset to load. Can be "pancreas" or "dentate_gyrus"}
}
\value{
anndata object of the selected test dataset
}
\description{
Loads the specified test dataset
}
\examples{
test_dataset <- load_test_dataset("pancreas")

}
