% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIERCE.R
\name{perform_preprocessing}
\alias{perform_preprocessing}
\title{perform_preprocessing}
\usage{
perform_preprocessing(
  adata,
  project_dir = "./Velocity_of_the_entropy_pipeline",
  min_genes = NULL,
  min_cells = NULL,
  compute_QC_metrics = TRUE,
  compute_MT_fraction = TRUE,
  MT_prefix = "MT-",
  UMI_lower_thr = NULL,
  UMI_upper_thr = NULL,
  Gene_lower_thr = NULL,
  Gene_upper_thr = NULL,
  MTf_lower_thr = NULL,
  MTf_upper_thr = NULL,
  do_normalization = TRUE,
  target_sum_for_norm = 10000,
  do_log_transform = TRUE,
  cell_cycle_scoring = FALSE,
  s_genes = NULL,
  g2m_genes = NULL,
  find_HVGs = TRUE,
  min_mean_HVGs = 0.0125,
  max_mean_HVGs = 3,
  min_disp_HVGs = 0.5,
  max_disp_HVGs = Inf,
  n_top_HVGs = NULL,
  flavor_HVGs = "seurat",
  subset_HVGs = FALSE,
  perform_PCA = TRUE,
  plot_PCA_heatmap = TRUE,
  scale_data = TRUE,
  perform_regression = FALSE,
  regress_vars = NULL,
  compute_neighbor_graph = TRUE,
  n_neighbors = 30,
  n_pcs = NULL,
  perform_clustering = TRUE,
  perform_UMAP = TRUE,
  color_as = NULL,
  lab_order = NULL,
  palette = NULL,
  legend_loc = "right margin",
  alpha = 1,
  add_outline = FALSE,
  find_DEGs = TRUE,
  groups_for_DEGs = NULL,
  method_for_DEGs = "wilcoxon",
  DEGs_to_show = 25,
  adata_copy = FALSE
)
}
\arguments{
\item{adata}{anndata object}

\item{project_dir}{name of the directory containing the results of the main FIERCE analysis (including the path). If it does not exist, it will be created. The default name is "./Velocity_of_the_entropy_pipeline". The plots of the Scanpy pipeline will be saved in the "scanpy_preprocessing" sub-directory}

\item{min_genes}{minimum number of expressed genes required for a cell to be retained in the anndata object}

\item{min_cells}{minimum number of expressing cells required for a gene to be retained in the anndata object}

\item{compute_QC_metrics}{boolean; whether to compute quality metrics on cells. TRUE by default}

\item{compute_MT_fraction}{boolean; whether to include the mitochondrial fraction in quality metrics. TRUE by default}

\item{MT_prefix}{prefix of mitochondrial genes. "MT-" (human) by default}

\item{UMI_lower_thr}{minimum number of total UMI counts required for a cell to be retained in the anndata object}

\item{UMI_upper_thr}{maximum number of total UMI counts allowed for a cell to be retained in the anndata object}

\item{Gene_lower_thr}{minimum number of expressed genes (by counts) required for a cell to be retained in the anndata object}

\item{Gene_upper_thr}{maximum number of expressed genes (by counts) allowed for a cell to be retained in the anndata object}

\item{MTf_lower_thr}{minimum mitochondrial fraction required for a cell to be retained in the anndata object}

\item{MTf_upper_thr}{maximum mitochondrial fraction allowed for a cell to be retained in the anndata object}

\item{do_normalization}{boolean; whether to perform library size normalization on cells. TRUE by default}

\item{target_sum_for_norm}{scale factor for library size normalization. 1e4 by default}

\item{do_log_transform}{boolean; whether to log-transform the expression matrix after normalization. TRUE by default}

\item{cell_cycle_scoring}{boolean; whether to perform cell cycle scoring on cells. FALSE by default. If TRUE, cell cycle genes must be provided by the user}

\item{s_genes}{character vector containing the S phase genes. Only used if cell_cycle_scoring=TRUE}

\item{g2m_genes}{character vector containing the G2M phase genes. Only used if cell_cycle_scoring=TRUE}

\item{find_HVGs}{boolean; whether to identify highly variable genes. TRUE by default}

\item{min_mean_HVGs}{lower cutoff for the mean of highly variable genes. Only used if find_HVGs=TRUE and n_top_HVGs=NULL. Default is 0.0125}

\item{max_mean_HVGs}{upper cutoff for the mean of highly variable genes. Only used if find_HVGs=TRUE and n_top_HVGs=NULL. Default is 3}

\item{min_disp_HVGs}{lower cutoff for the normalized dispersion of highly variable genes. Only used if find_HVGs=TRUE and n_top_HVGs=NULL. Default is 0.5}

\item{max_disp_HVGs}{upper cutoff for the normalized dispersion of highly variable genes. Only used if find_HVGs=TRUE and n_top_HVGs=NULL. Default is Inf}

\item{n_top_HVGs}{number of top variable genes to keep. Overrides all cutoffs specified for mean and normalized dispersion. Mandatory if flavor_HVGs="seurat_v3"}

\item{flavor_HVGs}{flavor for highly variable genes computation. Choose between "seurat", "cell_ranger" or "seurat_v3". Only used if find_HVGs=TRUE. Default is "seurat"}

\item{subset_HVGs}{boolean; whether to subset the anndata object to retain only the highly variable genes. Only used if find_HVGs=TRUE. Default is FALSE}

\item{perform_PCA}{boolean; whether to perform PCA. Default is TRUE}

\item{plot_PCA_heatmap}{boolean; whether to plot the heatmap of the expression of the most significant genes for each PC. Default is TRUE}

\item{scale_data}{boolean; whether to perform data scaling before PCA. Only used if perform_PCA=TRUE. Default is TRUE. WARNING: scaling is performed just for PCA, the expression matrix in the "X" slot will remain unscaled}

\item{perform_regression}{boolean; whether to perform a regression procedure during data scaling. Only used if perform_PCA=TRUE and scale_data=TRUE. Default is FALSE}

\item{regress_vars}{character vector containing the names of the numerical variables in the "obs" slot to include in the (optional) regression procedure. NULL by default}

\item{compute_neighbor_graph}{boolean; whether to compute the nearest neighbor graph on the PC space. Default is TRUE}

\item{n_neighbors}{number of nearest neighbors to compute for each cell in the nearest neighbor graph. Only used if compute_neighbor_graph=TRUE. Default is 30 neighbors}

\item{n_pcs}{number of PCs to consider for the nearest neighbor graph computation. Only used if compute_neighbor_graph=TRUE. Default (NULL) is all the computed PCs}

\item{perform_clustering}{boolean; whether to perform clustering analysis ("Leiden" algorithm). Default is TRUE. WARNING: a nearest neighbor graph must be computed first}

\item{perform_UMAP}{boolean; whether to compute UMAP embedding. Default is TRUE. WARNING: a nearest neighbor graph must be computed first}

\item{color_as}{character vector containing the names of the cell annotations in the "obs" slot to visualize on the PCA and UMAP plots. If present, the results of the clustering analysis will be visualized by default}

\item{lab_order}{optional; list containing, for each cell annotation specified in "color_as", a character vector specifying the order by which the respective labels should be printed in the legends of PCA and UMAP plots. If it is not necessary to specify a particular order for a specific annotation, just write NULL. If it is not necessary to specify any order for any annotation, do not change the default value of this parameter}

\item{palette}{optional; list containing, for each cell annotation specified in "color_as", a character vector specifying the colors of the respective labels. If it is not necessary to specify any color for a specific annotation, just write NULL; in this case, a default color palette will be used. If it is not necessary to specify any color for any annotation, do not change the default value of this parameter; in this case, annotations will be colored according either to a default palette, or to the colors that are already stored in the anndata object ("uns" slot)}

\item{legend_loc}{localization of the legend on the PCA and UMAP plots. Default is "right margin"}

\item{alpha}{transparency of the dots on the PCA and UMAP plots, ranging from 0 (fully transparent) to 1 (fully opaque). Default is 1}

\item{add_outline}{whether to add an outline around groups of dots on the PCA and UMAP plots. Default is FALSE}

\item{find_DEGs}{boolean; whether to find the differentially expressed genes (DEGs) between the groups defined in groups_for_DEGs. Default is TRUE}

\item{groups_for_DEGs}{name of the cell annotation in the "obs" slot to use as grouping variable for DEGs analysis. If present, the results of the clustering analysis will be used. If clustering analysis has not been performed, and the user does not specify an alternative grouping variable, the DEGs analysis will be skipped}

\item{method_for_DEGs}{method to use for DEGs analysis. Choose between "logreg", "t-test", "wilcoxon" or "t-test_overestim_var". Default is "wilcoxon"}

\item{DEGs_to_show}{number of DEGs to plot. Default is 25}

\item{adata_copy}{boolean; if TRUE, create a new anndata object with the results of the analysis; if FALSE, update the provided anndata object. Default is FALSE}
}
\value{
If adata_copy=TRUE, a new anndata object with the results of the analysis will be returned. Otherwise (default), the provided anndata object will be updated with the results of the analysis
}
\description{
Performs a full Scanpy analysis (optional) pipeline on the expression matrix (slot "X") of the provided anndata object, from quality checks to clustering. By tuning the appropriate parameters, the user is free to personalize the analysis and perform only the desired operations
}
\examples{
#Perform a complete analysis with default values

perform_preprocessing(adata)

#Perform a complete analysis after setting thresholds for QC metrics

perform_preprocessing(adata, min_genes=200, min_cells=3, UMI_lower_thr=1000, UMI_upper_thr=20000, Gene_lower_thr=500, Gene_upper_thr=6500, MTf_upper_thr=0.2)

#Perform a partial analysis on an anndata object that already contains pre-computed PCs: perform only clustering, UMAP and DEGs, skip all previous steps. Create a new anndata object with the results of the analysis

adata_preprocessed <- perform_preprocessing(adata, compute_QC_metrics=FALSE, do_normalization=FALSE, do_log_transform=FALSE, find_HVGs=FALSE, perform_PCA=FALSE, adata_copy=TRUE)

}
