% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIERCE.R
\name{plot_entropy_results}
\alias{plot_entropy_results}
\title{plot_entropy_results}
\usage{
plot_entropy_results(
  adata,
  project_dir = "./Velocity_of_the_entropy_pipeline",
  phenotype_annotation = NULL,
  phenotype_order = NULL,
  phenotype_colors = NULL,
  horizontal = FALSE
)
}
\arguments{
\item{adata}{anndata object containing the results of "compute_signaling_entropy"}

\item{project_dir}{name of the directory containing the results of the main FIERCE analysis (including the path). If it does not exist, it will be created. The default name is "./Velocity_of_the_entropy_pipeline". The plots will be saved in the "signaling_entropy_plots" sub-directory}

\item{phenotype_annotation}{character vector containing the names of the cell annotations in the "obs" slot to visualize on the boxplot of the total entropies and on the dotplot of the potency states. Any number of columns can be specified}

\item{phenotype_order}{optional; list containing, for each phenotype annotation specified in phenotype_annotation, a character vector specifying the order by which the respective phenotype labels should be plotted. If it is not necessary to specify a particular order for a specific annotation, just write NULL. If it is not necessary to specify any order for any annotation, do not change the default value of this parameter}

\item{phenotype_colors}{optional; list containing, for each phenotype annotation specified in phenotype_annotation, a character vector specifying the colors for the respective phenotype labels. If it is not necessary to specify any color for a specific annotation, just write NULL; in this case, a default color palette will be used. If it is not necessary to specify any color for any annotation, do not change the default value of this parameter; in this case, annotations will be colored according either to a default palette, or to the colors that are already stored in the anndata object ("uns" slot)}

\item{horizontal}{whether to plot cell annotations on the horizontal axis on the boxplots. FALSE by default}
}
\value{
No objects are returned, the plots will be saved in the directory containing the results of the main FIERCE analysis ("signaling_entropy_plots" sub-directory)
}
\description{
Plots the results of "compute_signaling_entropy". Saves several plots that allow a visual check of the results of the SCENT analysis, namely: a boxplot of total entropies per phenotype label (for one or more phenitype annotations chosen by the user), a boxplot of the total number of expressed genes per phenotype, a boxplot of the mean expression level per phenotype, a boxplot of the median expression level per phenotype, a scatterplot of total entropies VS total number of expressed genes, a scatterplot of total entropies VS mean expression level, a scatterplot of total entropies VS median expression level, and a dotplot of the abundancies of potency states within each phenotype label. An ANOVA test is performed for each phenotype annotation used for the boxplots. Pearson correlations between total entropies, total number of expressed genes, mean and median gene expression are also computed and printed on screen
}
\examples{
In this example, the clusters, the cell types and the cell cycle phase are chosen as phenotype annotations. A specific label order and a specific color for each label are specified for the cell cycle phase only

plot_entropy_results(adata_velocity_entropy, phenotype_annotation=c("clusters", "cell_types", "phase"), phenotype_order=list(NULL, NULL, c("G1","S","G2M")), phenotype_colors=list(NULL, NULL, c("green", "red", "blue")))

}
