% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIERCE.R
\name{plot_signature_statistics}
\alias{plot_signature_statistics}
\title{plot_signature_statistics}
\usage{
plot_signature_statistics(
  adata,
  signature = NULL,
  signature_name = NULL,
  project_dir = "./Velocity_of_the_entropy_pipeline",
  use_raw = FALSE,
  embedding_basis = "umap_entropy",
  n_neighbors = NULL,
  n_pcs = NULL,
  color_as = NULL,
  lab_order = NULL,
  palette = NULL,
  n_genes = NULL,
  legend_loc = "right"
)
}
\arguments{
\item{adata}{anndata object containing the results of "compute_velocity" and "compute_signaling_entropy"}

\item{signature}{optional character vector containing the gene signature provided by the user}

\item{signature_name}{if a gene signature is provided by the user, this name will be reported in the output file names}

\item{project_dir}{name of the directory containing the results of the main FIERCE analysis (including the path). If it does not exist, it will be created. The default name is "./Velocity_of_the_entropy_pipeline". The plots will be saved in the "signature_statistics_plots" sub-directory}

\item{use_raw}{boolean; whether RNA velocities and signaling entropies have been computed from the normalized counts instead of the first order moments (see documentation for "compute_velocity" and "compute_signaling_entropy" functions). This will affect the data that will be plotted as expression levels. Default is FALSE}

\item{embedding_basis}{name of the embedding to use for plotting the expression level, the partial entropy level, the RNA velocity and the velocity of the entropy. Default is "umap_entropy"}

\item{n_neighbors}{if "umap_entropy" is chosen as cell embedding for the scatterplots, this parameter, in combination with "n_pcs", specifies which embedding will be used among those that have been produced by the "compute_entropy_UMAP" function (stored in the "obsm" slot). If NULL (default), the first entropy-derived UMAP embedding stored in the "obsm" slot will be used}

\item{n_pcs}{if "umap_entropy" is chosen as cell embedding for the scatterplots, this parameter, in combination with "n_neighbors", specifies which embedding will be used among those that have been produced by the "compute_entropy_UMAP" function (stored in the "obsm" slot). If NULL (default), the first entropy-derived UMAP embedding sotred in the "obsm" slot will be used}

\item{color_as}{character vector containing the names of the cell annotations in the "obs" slot to visualize on the phase portraits. A distinct pdf file will be produced for each annotation}

\item{lab_order}{optional; list containing, for each cell annotation specified in "color_as", a character vector specifying the order by which the respective labels should be printed in the legends of the phase portraits. If it is not necessary to specify a particular order for a specific annotation, just write NULL. If it is not necessary to specify any order for any annotation, do not change the default value of this parameter}

\item{palette}{optional; list containing, for each cell annotation specified in "color_as", a character vector specifying the colors of the respective labels. If it is not necessary to specify any color for a specific annotation, just write NULL; in this case, a default color palette will be used. If it is not necessary to specify any color for any annotation, do not change the default value of this parameter; in this case, annotations will be colored according either to a default palette, or to the colors that are already stored in the anndata object ("uns" slot)}

\item{n_genes}{number of top-likelihood genes that will be shown in the plots. By default (NULL), the 20 top genes will be plotted if the user does not provide any custom signature, otherwise all the genes of the signature that passed the likelihood threshold will be plotted}

\item{legend_loc}{localization of the legend on the plots. Default is "right"}
}
\value{
No objects are returned, the plots will be saved in the directory containing the results of the main FIERCE analysis ("signature_statistics_plots" sub-directory)
}
\description{
Plots the phase portraits, the expression level, the partial entropy level, the RNA velocity and the velocity of the entropy of the genes that achieved the top likelihood score in the velocity dynamical model. Alternatively, the user can provide their own gene signature, that will be filtered to include only the subset of genes that successfully fitted the dynamical model. In any case, the genes will be plotted in decreasing order according to their likelihood score
}
\examples{
#Plot the phase portraits, the expression level, the partial entropy level, the RNA velocity and the velocity of the entropy of the top 20 likelihood genes. The first entropy-derived UMAP embedding stored in the anndata object is used. In this example, the clusters, the cell types and the cell cycle phase are visualized on the phase portraits. A specific label order and a specific color for each label is specified for the cell cycle phase only.

plot_signature_statistics(adata_velocity_entropy, color_as=c("clusters", "cell_types", "phase"), lab_order=list(NULL, NULL, c("G1","S","G2M")), palette=list(NULL, NULL, c("green", "red", "blue")))

}
